backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
django-nested-inline==0.4.6
drf-writable-nested==0.7.0